requirements.txt
pytest-playwright>=0.7.0
allure-pytest>=2.14.0
pytest>=8.3.5
jsonmodels==2.7.0
loguru==0.7.3